Make '/' redirect to '/client/dashboard' if user is authenticated
I created a new annotation @TheClient which can be used instead of the
@AuthentificationPrincipal annotation. This allows for less spring
security code in our web layer. This annotation is then used in
deciding if the user is logged in or not.

For consistency all other routes now use this annotation as well.

This broke some (already fixed) tests because various routes expected to
get a Principal but now get a CurrentClient, which hides Spring
Security in tests as well.